Fix speakerphone in JB 4.1 (1/2)
On Android 4.1 the speakerphone does not work correctly - the volume remains low and the audio server often crashes and causes a hot reboot of the device.
The problem is mainly caused by the p500 audio libraries which are not needed and just give problems. This part of the commit brings the libaudio inline with Android 4.2's, to avoid
further problems.

Change-Id: I6748aafe3f6103909eb8f86bce1b5a4d2f5746cc